chore(ci): Tighten uplift criteria
As discussed in the Drivers channel this week, we want to limit
uplifts to Beta, Release, and ESR as much as possible. This change
aligns the Beta criteria with the existing Release/ESR criteria.

The uplift template is also dropped as it can become out of
date with the official template that is used in github.